Summery
SampleUnits change to show only years and for Controller, for PeriodField setEditable(false)
And change Controller names to right ones
